Enabled support for _DBFileLock state locking

+++ internal/db/sqlite/sqlite_driver.go
@@ -3,6 +3,7 @@ package database
 import (
 	"GoSally/internal/logger"
 	"database/sql"
+	"errors"
 	_ "modernc.org/sqlite"
 	"os"
 	"path/filepath"
@@ -12,6 +13,13 @@ import (
 type SQLiteDriver struct {
 	db   *sql.DB
 	_src string
+
+	/*
+	 * Locks the database interface.
+	 * _DBFileLock true = file is open
+	 * _DBFileLock false = file is closed
+	 */
+	_DBFileLock bool
 }
 
 func ensureDBPath(src string) error {
@@ -20,9 +28,12 @@ func ensureDBPath(src string) error {
 }
 
 func (s *SQLiteDriver) OpenDB(src string) error {
-	var err error
+	if s._DBFileLock {
+		return errors.New("DB is already unlocked (opened)")
+	}
+
 	s._src = src
-	err = ensureDBPath(s._src)
+	err := ensureDBPath(s._src)
 	if err != nil {
 		logger.DatabaseLog.Error("Error creating database directory", "db_source", s._src, "err", err)
 	}
@@ -44,10 +55,14 @@ func (s *SQLiteDriver) OpenDB(src string) error {
 		logger.DatabaseLog.Error("Error executing command in database", "db_source", s._src, "err", err)
 		return err
 	}
+	s._DBFileLock = true
 	return nil
 }
 
 func (s *SQLiteDriver) CloseDB() error {
+	if !s._DBFileLock {
+		return errors.New("DB is already locked (closed)")
+	}
 	logger.DatabaseLog.Debug("Closing a database", "db_source", s._src)
 	err := s.db.Close()
 	if err != nil {
@@ -55,10 +70,14 @@ func (s *SQLiteDriver) CloseDB() error {
 		return err
 	}
 	logger.DatabaseLog.Debug("The database is closed", "db_source", s._src)
+	s._DBFileLock = false
 	return nil
 }
 
 func (s *SQLiteDriver) InitSession(id string, data []byte) error {
+	if !s._DBFileLock {
+		return errors.New("DB is locked (closed)")
+	}
 	_, err := s.db.Exec("INSERT INTO sessions(session_id, data) VALUES(?, ?)", id, data)
 	if err != nil {
 		logger.DatabaseLog.Error("Error executing command in database", "db_source", s._src, "err", err)
@@ -68,6 +87,9 @@ func (s *SQLiteDriver) InitSession(id string, data []byte) error {
 }
 
 func (s *SQLiteDriver) QuerySession(id string) (data []byte, err error) {
+	if !s._DBFileLock {
+		return nil, errors.New("DB is locked (closed)")
+	}
 	row := s.db.QueryRow("SELECT data FROM sessions WHERE session_id=?", id)
 	err = row.Scan(&data)
 	if err != nil {
@@ -78,6 +100,9 @@ func (s *SQLiteDriver) QuerySession(id string) (data []byte, err error) {
 }
 
 func (s *SQLiteDriver) CloseSession(id string) error {
+	if !s._DBFileLock {
+		return errors.New("DB is locked (closed)")
+	}
 	_, err := s.db.Exec("DELETE FROM sessions WHERE session_id=?", id)
 	if err != nil {
 		logger.DatabaseLog.Error("Error executing command in database", "db_source", s._src, "err", err)
